requirements.txt
certifi==2019.11.28
chardet==3.0.4
Click==7.0
emnist==0.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
requests==2.22.0
six==1.14.0
tqdm==4.42.0
urllib3==1.25.8
Werkzeug==0.16.1
